WORKING I/Ogit add .!

+++ kernel/include/ios/inode.h
@@ -12,14 +12,14 @@ typedef struct inode {
     uint16_t major, minor;
 
     // disk files only
-    uint16_t num; // for the use of the filesystem driver
+    uint32_t num; // for the use of the filesystem driver
     size_t size;
     int blkdev;
     blknum_t (*bmap)(struct inode*, blknum_t); // maps file blocks to disk blocks
 } inode_t;
 
 void creatid(inode_t* ip, uint16_t major, uint16_t minor);
-void creatif(inode_t* ip, uint16_t num, size_t size, uint16_t blkdev, blknum_t (*bmap)(struct inode*, blknum_t));
+void creatif(inode_t* ip, uint32_t num, size_t size, uint16_t blkdev, blknum_t (*bmap)(struct inode*, blknum_t));
 int readi(inode_t* ip, uint8_t* dst, size_t off, size_t n);
 
 #define I_DEVICE 0
